feat: Implement offline sync functionality with queue management for various operations

+++ client/scripts/Core/controllers/DepartmentHandler.js
@@ -1,3 +1,4 @@
+import { syncQueueHandler } from "../startup.js";
 class DepartmentHandler {
   constructor(deptRepo, userState) {
     this.deptRepo = deptRepo;
@@ -26,6 +27,18 @@ class DepartmentHandler {
   async EditDepartment(id, name, description) {
     if (!this.user || !this.user.data || !this.user.data.user || this.user.data.user.role !== "HR") return { ok: false, data: "Please login first" }
     try {
+      const sync = await syncQueueHandler.AddItemToQueue("departments", "edit", {
+        id,
+        name,
+        description
+      });
+      console.log("Sync Queue Response: ", sync);
+      if (!sync.ok) {
+        return {
+          ok: false,
+          data: "Unable to store in sync queue try to get online cause offline sync queue to gayi: " + sync.data
+        }
+      }
       const { ok, data } = await this.deptRepo.EditDepartment(id, name, description);
       return {
         ok,
@@ -45,6 +58,16 @@ class DepartmentHandler {
   async DeleteDepartment(id) {
     if (!this.user || !this.user.data || !this.user.data.user || this.user.data.user.role !== "HR") return { ok: false, data: "Please login first" }
     try {
+      const sync = await syncQueueHandler.AddItemToQueue("departments", "delete", {
+        id
+      });
+      console.log("Sync Queue Response: ", sync);
+      if (!sync.ok) {
+        return {
+          ok: false,
+          data: "Unable to store in sync queue try to get online cause offline sync queue to gayi: " + sync.data
+        }
+      }
       const { ok, data } = await this.deptRepo.DeleteDepartment(id);
       return {
         ok,
