+++ client/scripts/Core/controllers/SalaryHandler.js
@@ -1,3 +1,4 @@
+import {syncQueueHandler} from "../startup.js";
 class SalaryHandler {
     constructor(salaryRepo, authState) {
         this.salaryRepo = salaryRepo;
@@ -13,6 +14,22 @@ class SalaryHandler {
             }
         }
         try {
+            const startSync = await syncQueueHandler.AddItemToQueue("salaries", "create", {
+                userId,
+                base,
+                hra,
+                lta,
+                userFirstName,
+                userLastName,
+                userDepartment
+            });
+            console.log("Start Sync Response: ", startSync);
+            if (!startSync.ok) {
+                return {
+                    ok: false,
+                    data: "Unable to store in sync queue try to get online cause offline sync queue to gayi: " + startSync.data
+                }
+            }
             const data = await this.salaryRepo.Create(base, hra, lta, userId, userFirstName, userLastName, userDepartment);
             return {
                 ok: true,
@@ -71,6 +88,16 @@ class SalaryHandler {
             }
         }
         try {
+            const startSync = await syncQueueHandler.AddItemToQueue("salaries", "delete", {
+                id
+            });
+            console.log("Start Sync Response: ", startSync);
+            if (!startSync.ok) {
+                return {
+                    ok: false,
+                    data: "Unable to store in sync queue try to get online cause offline sync queue to gayi: " + startSync.data
+                }
+            }
             const data = await this.salaryRepo.DeleteSalary(id);
             return {
                 ok: true,
@@ -92,6 +119,19 @@ class SalaryHandler {
             }
         }
         try {
+            const startSync = await syncQueueHandler.AddItemToQueue("salaries", "edit", {
+                id,
+                base,
+                hra,
+                lta
+            });
+            console.log("Start Sync Response: ", startSync);
+            if (!startSync.ok) {
+                return {
+                    ok: false,
+                    data: "Unable to store in sync queue try to get online cause offline sync queue to gayi: " + startSync.data
+                }
+            }
             const data = await this.salaryRepo.EditSalary(id, base, hra, lta);
             return {
                 ok: true,
