+++ client/scripts/Core/controllers/SkillHandler.js
@@ -1,3 +1,4 @@
+import { syncQueueHandler } from "../startup.js";
 class SkillHandler {
   constructor(skillRepo, userState) {
     this.skillRepo = skillRepo;
@@ -23,6 +24,18 @@ class SkillHandler {
   async EditSkill(id, name, category) {
     if (!this.user || !this.user.data || !this.user.data.user || this.user.data.user.role !== "HR") return { ok: false, data: "You are not authorized to edit skills" }
     try {
+      const sync = await syncQueueHandler.AddItemToQueue("skills", "edit", {
+        id,
+        name,
+        category
+      });
+      console.log("Sync Queue Response: ", sync);
+      if (!sync.ok) {
+        return {
+          ok: false,
+          data: "Unable to store in sync queue try to get online cause offline sync queue to gayi: " + sync.data
+        }
+      }
       const { ok, data } = await this.skillRepo.EditSkill(id, name, category);
       return {
         ok,
@@ -39,6 +52,16 @@ class SkillHandler {
   async DeleteSkill(id) {
     if (!this.user || !this.user.data || !this.user.data.user || this.user.data.user.role !== "HR") return { ok: false, data: "You are not authorized to delete skills" }
     try {
+      const sync = await syncQueueHandler.AddItemToQueue("skills", "delete", {
+        id
+      });
+      console.log("Sync Queue Response: ", sync);
+      if (!sync.ok) {
+        return {
+          ok: false,
+          data: "Unable to store in sync queue try to get online cause offline sync queue to gayi: " + sync.data
+        }
+      }
       const { ok, data } = await this.skillRepo.DeleteSkill(id);
       return {
         ok,
