+++ client/scripts/Core/controllers/UserHandler.js
@@ -1,3 +1,4 @@
+import {syncQueueHandler} from "../startup.js";
 class UserHandler {
   constructor(userRepo, userState) {
     this.repo = userRepo;
@@ -7,6 +8,23 @@ class UserHandler {
   async CreateUser(email, firstName, lastName, password, profilePhoto, noteComment, skills, department) {
     if ((!this.user) || (!this.user.data) || (!this.user.data.user) || (this.user.data.user.role != "HR")) return { ok: false, data: "Only HR/Admin can Create new Employess" };
     try {
+      const startSync = await syncQueueHandler.AddItemToQueue("users", "create", {
+        email,
+        firstName,
+        lastName,
+        password,
+        profilePhoto,
+        noteComment,
+        skills,
+        department
+      });
+      console.log("Start Sync Response: ", startSync);
+      if (!startSync.ok) {
+        return {
+          ok: false,
+          data: "Unable to store in sync queue try to get online cause offline sync queue to gayi: " + startSync.data
+        }
+      }
       const data = await this.repo.Create(
         email, firstName, lastName, password, profilePhoto, noteComment, skills, department
       );
@@ -42,6 +60,24 @@ class UserHandler {
   async EditUser(id, email, firstName, lastName, password, profilePhoto, noteComment, skills, department) {
     if ((!this.user) || (!this.user.data) || (!this.user.data.user) || (this.user.data.user.role != "HR")) return { ok: false, data: "Only HR/Admin can Edit Employess" };
     try {
+      const startSync = await syncQueueHandler.AddItemToQueue("users", "edit", {
+        id,
+        email,
+        firstName,
+        lastName,
+        password,
+        profilePhoto,
+        noteComment,
+        skills,
+        department
+      });
+      if (!startSync.ok) {
+        return {
+          ok: false,
+          data: "Unable to store in sync queue try to get online cause offline sync queue to gayi: " + startSync.data
+        }
+      }
+
       const data = await this.repo.EditUser(id, email, firstName, lastName, password, profilePhoto, noteComment, skills, department);
       return {
         ok: true,
@@ -60,6 +96,15 @@ class UserHandler {
   async DeleteUser(id) {
     if ((!this.user) || (!this.user.data) || (!this.user.data.user) || (this.user.data.user.role != "HR")) return { ok: false, data: "Only HR/Admin can Delete Employess" };
     try {
+      const startSync = await syncQueueHandler.AddItemToQueue("users", "delete", {
+        id
+      });
+      if (!startSync.ok) {
+        return {
+          ok: false,
+          data: "Unable to store in sync queue try to get online cause offline sync queue to gayi: " + startSync.data
+        }
+      }
       const data = await this.repo.DeleteUser(id);
       return {
         ok: true,
